+++ internal/storage/sqlstore.go
@@ -8,8 +8,8 @@ import (
 	"strings"
 	"time"
 
-	"codeberg.org/pivpav/glow/internal/article"
-	"codeberg.org/pivpav/glow/internal/config"
+	"github.com/piv-pav/glow/internal/article"
+	"github.com/piv-pav/glow/internal/config"
 	_ "github.com/rqlite/gorqlite/stdlib"
 	_ "modernc.org/sqlite"
 )