+++ tools/delete.go
@@ -3,8 +3,8 @@ package tools
 import (
 	"fmt"
 
-	"codeberg.org/pivpav/glow/internal/article"
-	"codeberg.org/pivpav/glow/internal/storage"
+	"github.com/piv-pav/glow/internal/article"
+	"github.com/piv-pav/glow/internal/storage"
 	"github.com/spf13/cobra"
 )