+++ tools/move.go
@@ -3,7 +3,7 @@ package tools
 import (
 	"fmt"
 
-	"codeberg.org/pivpav/glow/internal/storage"
+	"github.com/piv-pav/glow/internal/storage"
 	"github.com/spf13/cobra"
 )